draft: split running tests from getting access commands
