+++ bin/run_tests.mjs
@@ -58,42 +58,3 @@ k6_run(tester,
     {commit: commit, cluster: "upstream", test: "create_projects.mjs", Projects: PROJECT_COUNT},
     "k6/create_projects.js", true
 )
-
-const testerAddresses = getAppAddressesFor(tester)
-
-// Output access details
-console.log("*** ACCESS DETAILS")
-console.log()
-
-console.log(`*** UPSTREAM CLUSTER`)
-console.log(`    Rancher UI: ${rancherURL} (admin/${ADMIN_PASSWORD})`)
-
-console.log(`    Kubernetes API:`)
-console.log(`export KUBECONFIG=${q(upstream["kubeconfig"])}`)
-console.log(`kubectl config use-context ${q(upstream["context"])}`)
-for (const [node, command] of Object.entries(upstream["node_access_commands"])) {
-    console.log(`    Node ${node}: ${q(command)}`)
-}
-console.log()
-
-for (const [name, downstream] of downstreams) {
-    console.log(`*** ${name.toUpperCase()} CLUSTER`)
-    console.log(`    Kubernetes API:`)
-    console.log(`export KUBECONFIG=${q(downstream["kubeconfig"])}`)
-    console.log(`kubectl config use-context ${q(downstream["context"])}`)
-    for (const [node, command] of Object.entries(downstream["node_access_commands"])) {
-        console.log(`    Node ${node}: ${q(command)}`)
-    }
-    console.log()
-}
-
-console.log(`*** TESTER CLUSTER`)
-const grafanaURL = testerAddresses.localNetwork.httpURL
-console.log(`    Grafana UI: ${grafanaURL}/grafana/d/a1508c35-b2e6-47f4-94ab-fec400d1c243/test-results?orgId=1&refresh=5s&from=now-30m&to=now (admin/${ADMIN_PASSWORD})`)
-console.log(`    Kubernetes API:`)
-console.log(`export KUBECONFIG=${q(tester["kubeconfig"])}`)
-console.log(`kubectl config use-context ${q(tester["context"])}`)
-for (const [node, command] of Object.entries(tester["node_access_commands"])) {
-    console.log(`    Node ${node}: ${q(command)}`)
-}
-console.log()
